Restructure package
+ Move jsonfeed from module into __init__.py. Ugly, but it gets the
  default import dynamics I want.
+ Put feedparser converter in converters package without __init__.py.
  Clients must use the `from jsonfeed.converters import feedparser`
  pattern.
+ Add feedparser dependency to requirements.txt.
