+++ docs/index.html
@@ -42,9 +42,8 @@
         <div>
             <h2>Available Modules</h2>
             <ul>
-                    <li><a href="jsonfeed/converters.html">jsonfeed.converters</a></li>
+                    <li><a href="jsonfeed.html">jsonfeed</a></li>
                     <li><a href="jsonfeed/converters/feedparser.html">jsonfeed.converters.feedparser</a></li>
-                    <li><a href="jsonfeed/jsonfeed.html">jsonfeed.jsonfeed</a></li>
             </ul>
         </div>
     </nav>
