+++ docs/jsonfeed.html
@@ -216,7 +216,7 @@ jsonfeed    </h1>
 
 <h2 id="deprecations">Deprecations</h2>
 
-<p>See <a href="https://jsonfeed.org/">the spec</a> for an overview of deprecated JSON Feed fields. This project (especially the <code><a href="converters.html">converters</a></code> and the parsing functions) will stay backwards-compatible when possible, but using deprecated fields when constructing feeds is discouraged.</p>
+<p>See <a href="https://jsonfeed.org/">the spec</a> for an overview of deprecated JSON Feed fields. This project (especially the <code>converters</code> and the parsing functions) will stay backwards-compatible when possible, but using deprecated fields when constructing feeds is discouraged.</p>
 
 <h3 id="json-feed-11">JSON Feed 1.1</h3>
