--- jsonfeed/converters/__init__.py
@@ -1,1 +1,0 @@
-from .feedparser import *
