+++ requirements.txt
@@ -1,3 +1,5 @@
+feedparser==6.0.2
+
 # Development dependencies
 pytest>=6.2.2
 flake8>=3.9.0
